+++ Server/Persistence/StandardSaveStrategy.cs
@@ -80,14 +80,14 @@ namespace Server
                 bin = new AsyncWriter(World.MobileDataPath, true);
             }
 
-            idx.Write((int)mobiles.Count);
+            idx.Write(mobiles.Count);
             foreach (Mobile m in mobiles.Values)
             {
                 long start = bin.Position;
 
-                idx.Write((int)m.m_TypeRef);
-                idx.Write((int)m.Serial);
-                idx.Write((long)start);
+                idx.Write(m.m_TypeRef);
+                idx.Write(m.Serial);
+                idx.Write(start);
 
                 m.Serialize(bin);
 
@@ -101,7 +101,7 @@ namespace Server
                 m.FreeCache();
             }
 
-            tdb.Write((int)World.m_MobileTypes.Count);
+            tdb.Write(World.m_MobileTypes.Count);
 
             for (int i = 0; i < World.m_MobileTypes.Count; ++i)
                 tdb.Write(World.m_MobileTypes[i].FullName);
@@ -132,7 +132,7 @@ namespace Server
                 bin = new AsyncWriter(World.ItemDataPath, true);
             }
 
-            idx.Write((int)items.Count);
+            idx.Write(items.Count);
             foreach (Item item in items.Values)
             {
                 if (item.Decays && item.Parent == null && item.Map != Map.Internal && (item.LastMoved + item.DecayTime) <= DateTime.UtcNow)
@@ -142,9 +142,9 @@ namespace Server
 
                 long start = bin.Position;
 
-                idx.Write((int)item.m_TypeRef);
-                idx.Write((int)item.Serial);
-                idx.Write((long)start);
+                idx.Write(item.m_TypeRef);
+                idx.Write(item.Serial);
+                idx.Write(start);
 
                 item.Serialize(bin);
 
@@ -158,7 +158,7 @@ namespace Server
                 item.FreeCache();
             }
 
-            tdb.Write((int)World.m_ItemTypes.Count);
+            tdb.Write(World.m_ItemTypes.Count);
             for (int i = 0; i < World.m_ItemTypes.Count; ++i)
                 tdb.Write(World.m_ItemTypes[i].FullName);
 
@@ -183,14 +183,14 @@ namespace Server
                 bin = new AsyncWriter(World.GuildDataPath, true);
             }
 
-            idx.Write((int)BaseGuild.List.Count);
+            idx.Write(BaseGuild.List.Count);
             foreach (BaseGuild guild in BaseGuild.List.Values)
             {
                 long start = bin.Position;
 
-                idx.Write((int)0);//guilds have no typeid
-                idx.Write((int)guild.Id);
-                idx.Write((long)start);
+                idx.Write(0);//guilds have no typeid
+                idx.Write(guild.Id);
+                idx.Write(start);
 
                 guild.Serialize(bin);
